+++ src/builder.ts
@@ -8,8 +8,7 @@ import react from "@vitejs/plugin-react";
 import * as swc from "@swc/core";
 
 import type { Config } from "./config.js";
-import type { GetEntry, Prerenderer } from "./server.js";
-import { renderRSC, prefetcherRSC } from "./middleware/lib/rsc-handler.js";
+import { prerenderRSC } from "./middleware/lib/rsc-handler.js";
 
 // TODO we have duplicate code here and rscPrd.ts and rsc-handler*.ts
 
@@ -65,121 +64,12 @@ const rscAnalyzePlugin = (
   };
 };
 
-const prerender = async (
-  dir: string,
-  distPath: string,
-  publicPath: string,
-  distEntriesFile: string,
-  basePath: string,
-  publicIndexHtmlFile: string
-) => {
-  const { prerenderer, clientEntries } = await (import(
-    distEntriesFile
-  ) as Promise<{
-    getEntry: GetEntry;
-    prerenderer?: Prerenderer;
-    clientEntries?: Record<string, string>;
-  }>);
-
-  const getClientEntry = (id: string) => {
-    if (!clientEntries) {
-      throw new Error("Missing client entries");
-    }
-    const clientEntry =
-      clientEntries[id] ||
-      clientEntries[id.replace(/\.js$/, ".ts")] ||
-      clientEntries[id.replace(/\.js$/, ".tsx")];
-    if (!clientEntry) {
-      throw new Error("No client entry found");
-    }
-    return clientEntry;
-  };
-  const decodeId = (encodedId: string): [id: string, name: string] => {
-    let [id, name] = encodedId.split("#") as [string, string];
-    if (!id.startsWith("wakuwork/")) {
-      id = path.relative("file://" + encodeURI(path.join(dir, distPath)), id);
-      id = basePath + getClientEntry(decodeURI(id));
-    }
-    return [id, name];
-  };
-
-  if (prerenderer) {
-    const {
-      entryItems = [],
-      paths = [],
-      unstable_customCode = () => "",
-    } = await prerenderer();
-    await Promise.all(
-      Array.from(entryItems).map(async ([rscId, props]) => {
-        // FIXME we blindly expect JSON.stringify usage is deterministic
-        const serializedProps = JSON.stringify(props);
-        const searchParams = new URLSearchParams();
-        searchParams.set("props", serializedProps);
-        const destFile = path.join(
-          dir,
-          publicPath,
-          "RSC",
-          decodeURIComponent(rscId),
-          decodeURIComponent(`${searchParams}`)
-        );
-        fs.mkdirSync(path.dirname(destFile), { recursive: true });
-        await new Promise<void>((resolve, reject) => {
-          const stream = fs.createWriteStream(destFile);
-          stream.on("finish", resolve);
-          stream.on("error", reject);
-          renderRSC(
-            {
-              rscId,
-              props: props as any,
-            },
-            true
-          ).pipe(stream);
-        });
-      })
-    );
-
-    const publicIndexHtml = fs.readFileSync(publicIndexHtmlFile, {
-      encoding: "utf8",
-    });
-    for (const pathItem of paths) {
-      const code = await prefetcherRSC(pathItem, true);
-      const destFile = path.join(
-        dir,
-        publicPath,
-        pathItem,
-        pathItem.endsWith("/") ? "index.html" : ""
-      );
-      let data = "";
-      if (fs.existsSync(destFile)) {
-        data = fs.readFileSync(destFile, { encoding: "utf8" });
-      } else {
-        fs.mkdirSync(path.dirname(destFile), { recursive: true });
-        data = publicIndexHtml;
-      }
-      if (code) {
-        // HACK is this too naive to inject script code?
-        data = data.replace(/<\/body>/, `<script>${code}</script></body>`);
-      }
-      const code2 = unstable_customCode(pathItem, decodeId);
-      if (code2) {
-        data = data.replace(/<\/body>/, `<script>${code2}</script></body>`);
-      }
-      fs.writeFileSync(destFile, data, { encoding: "utf8" });
-    }
-  }
-};
-
 export async function runBuild(config: Config = {}) {
   const dir = path.resolve(config.build?.dir || ".");
   const basePath = config.build?.basePath || "/";
   const distPath = config.files?.dist || "dist";
   const publicPath = path.join(distPath, config.files?.public || "public");
   const indexHtmlFile = path.join(dir, config.files?.indexHtml || "index.html");
-  const publicIndexHtmlFile = path.join(
-    dir,
-    publicPath,
-    config.files?.indexHtml || "index.html"
-  );
   const distEntriesFile = path.join(
     dir,
     distPath,
@@ -302,17 +192,7 @@ export async function runBuild(config: Config = {}) {
     `export const clientEntries=${JSON.stringify(clientEntries)};`
   );
 
-  if (!"STILL WIP") {
-    // TODO still wip
-    await prerender(
-      dir,
-      distPath,
-      publicPath,
-      distEntriesFile,
-      basePath,
-      publicIndexHtmlFile
-    );
-  }
+  await prerenderRSC(true);
 
   const origPackageJson = require(path.join(dir, "package.json"));
   const packageJson = {
